# be/services/docs/audit-service/swagger.yaml
basePath: /api/v1
info:
  contact: {}
  description: API documentation for audit-service.
  title: Base Golang Audit Service API
  version: "1.0"
paths:
...
  /api-keys:
    get:
      parameters:
      - description: Client ID
        in: query
        name: client_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      parameters:
      - description: Create API key payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - API Keys
  /api-keys/{id}:
    delete:
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete API key
      tags:
      - API Keys
    get:
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get API key
      tags:
      - API Keys
    put:
      consumes:
      - application/json
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update API key payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update API key
      tags:
      - API Keys
  /api-keys/{id}/invalidate:
    post:
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Invalidate API key
      tags:
      - API Keys
  /audit-logs:
    get:
      parameters:
      - description: Client ID
        in: query
        name: client_id
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List audit logs by client
      tags:
      - Audit Logs
  /audit-logs/{logId}:
    get:
      parameters:
      - description: Log ID
        in: path
        name: logId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get audit log by ID
      tags:
      - Audit Logs
  /audit-logs/table/{tableName}:
    get:
      parameters:
      - description: Table name
        in: path
        name: tableName
        required: true
        type: string
      - description: Record ID
        in: query
        name: record_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List audit logs by table
      tags:
      - Audit Logs
...
  /dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get dashboard summary
      tags:
      - Dashboard
  /dashboard/asset-status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get dashboard asset status
      tags:
      - Dashboard
...
  /dashboard/device-status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get dashboard device status
      tags:
      - Dashboard
  /dashboard/insights:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get dashboard insights
      tags:
      - Dashboard
  /dashboard/problem-devices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get problem devices
      tags:
      - Dashboard
  /employees:
    get:
      parameters:
      - description: Client ID
        in: query
        name: client_id
        type: integer
      - description: Branch ID
        in: query
        name: branch_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List employees
      tags:
      - Employees
    post:
      consumes:
      - application/json
      parameters:
      - description: Create employee payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create employee
      tags:
      - Employees
  /employees/{id}:
    delete:
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete employee
      tags:
      - Employees
    get:
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get employee
      tags:
      - Employees
    put:
      consumes:
      - application/json
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update employee payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update employee
      tags:
      - Employees
  /employees/branch/{branchId}:
    get:
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List employees by branch
      tags:
      - Employees
  /employees/client/{clientId}:
    get:
      parameters:
      - description: Client ID
        in: path
        name: clientId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List employees by client
      tags:
      - Employees
...
  /reports:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List reports
      tags:
      - Reports
  /reports/assets/export:
    get:
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Export assets report
      tags:
      - Reports
  /reports/assignments/export:
    get:
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Export assignments report
      tags:
      - Reports
  /reports/devices/export:
    get:
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Export devices report
      tags:
      - Reports
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"